+++ .psscripts/build-functions.ps1
@@ -270,16 +270,17 @@ function Get-NetCoreSdkFromWeb ($version)
 
     $os  = if (Test-IsWindows) { "windows" } else { "linux" }
     $ext = if (Test-IsWindows) { ".zip" } else { ".tar.gz" }
+    $uri = "https://www.microsoft.com/net/download/thank-you/dotnet-sdk-$version-$os-x64-binaries"
+    Write-Host "Finding download link..."
 
-    $response = Invoke-WebRequest `
-                    -Uri "https://www.microsoft.com/net/download/thank-you/dotnet-sdk-$version-$os-x64-binaries" `
-                    -Method Get `
-                    -MaximumRedirection 0 `
+    $response = Invoke-WebRequest -Uri $uri
 
     $downloadLink =
         $response.Links `
             | Where-Object { $_.onclick -eq "recordManualDownload()" } `
             | Select-Object -Expand href
+
+    Write-Host "Creating temporary file..."
 
     $tempFile  = [System.IO.Path]::GetTempFileName() + $ext
 
@@ -308,8 +309,12 @@ function Install-NetCoreSdkFromArchive ($sdkArchivePath)
         Write-Host "Created folder '$env:DOTNET_INSTALL_DIR'."
         Expand-Archive -LiteralPath $sdkArchivePath -DestinationPath $env:DOTNET_INSTALL_DIR -Force
         Write-Host "Extracted '$sdkArchivePath' to folder '$env:DOTNET_INSTALL_DIR'."
-        $env:Path = "$env:DOTNET_INSTALL_DIR;$env:Path"
-        Write-Host "Added '$env:DOTNET_INSTALL_DIR' to the environment variables."
+        $updatedPath = "$env:DOTNET_INSTALL_DIR;$env:Path"
+        [Environment]::SetEnvironmentVariable("Path", $updatedPath, "Process")
+        [Environment]::SetEnvironmentVariable("Path", $updatedPath, "User")
+        [Environment]::SetEnvironmentVariable("Path", $updatedPath, "Machine")
+        Write-Host "Added '$dotnetInstallDir' to the Path environment variable:"
+        Write-Host $env:Path
     }
     else
     {
@@ -318,8 +323,12 @@ function Install-NetCoreSdkFromArchive ($sdkArchivePath)
         Write-Host "Created folder '$dotnetInstallDir'."
         Invoke-Cmd "tar -xf $sdkArchivePath -C $dotnetInstallDir"
         Write-Host "Extracted '$sdkArchivePath' to folder '$dotnetInstallDir'."
-        $env:PATH = "$env:PATH:$dotnetInstallDir"
-        Write-Host "Added '$dotnetInstallDir' to the environment variables."
+        $updatedPath = "$dotnetInstallDir`:$env:PATH"
+        [Environment]::SetEnvironmentVariable("PATH", $updatedPath, "Process")
+        [Environment]::SetEnvironmentVariable("PATH", $updatedPath, "User")
+        [Environment]::SetEnvironmentVariable("PATH", $updatedPath, "Machine")
+        Write-Host "Added '$dotnetInstallDir' to the PATH environment variable:"
+        Write-Host $env:PATH
     }
 }